+++ app/controllers/images_controller.rb
@@ -1,4 +1,5 @@
 class ImagesController < ApplicationController
+  before_action :authenticate_user
 
   def new
     @user = User.find(params[:user_id])
@@ -17,15 +18,19 @@ class ImagesController < ApplicationController
 
     @user = User.find(@image.user_id)
 
-    if @image.save
-      flash[:success] = "Congrats. Your image was added below."
+    if current_user.id == image.user_id || current_user.admin
+      if @image.save
+        flash[:success] = "Congrats. Your image was added below."
+      else
+        flash[:warning] = "Oops. Please try again."
+      end
+      # redirect_to "/familymembers/#{@user.id}"
+      respond_to do |format|
+        format.html { redirect_to "/familymembers/#{@user.id}" }
+        format.js { render 'create.js.erb' } #default behaviour is to run app/views/notes/create.js.erb file
+      end
     else
-      flash[:warning] = "Oops. Please try again."
-    end
-    # redirect_to "/familymembers/#{@user.id}"
-    respond_to do |format|
-      format.html { redirect_to "/familymembers/#{@user.id}" }
-      format.js { render 'create.js.erb' } #default behaviour is to run app/views/notes/create.js.erb file
+      redirect_to "/familymembers"
     end
   end
 
@@ -41,9 +46,14 @@ class ImagesController < ApplicationController
   def destroy
     image = Image.find(params[:id])
     user = User.find(image.user_id)
-    image.destroy
-    flash[:warning] = "You deleted your image."
-    redirect_to "/familymembers/#{user.id}"
+
+    if current_user.id == image.user_id || current_user.admin
+      image.destroy
+      flash[:warning] = "You deleted your image."
+      redirect_to "/familymembers/#{user.id}"
+    else
+      redirect_to "/familymembers"
+    end
   end
 
 end
