+++ app/controllers/status_updates_controller.rb
@@ -1,24 +1,26 @@
 class StatusUpdatesController < ApplicationController
+  before_action :authenticate_user!
 
   def index
     @title = "My Newsfeed"
-    if current_user
-      @status_updates = Group.find(current_user.group_id).status_updates.order('created_at DESC').limit(5)
 
-      @familymembers = User.where(group_id: current_user.group_id)
-      @familymembers.each do |familymember|
+    @status_updates = Group.find(current_user.group_id).status_updates.order('created_at DESC').limit(5)
+
+    @familymembers = User.where(group_id: current_user.group_id)
+
+    @familymembers.each do |familymember|
+      if familymember.birthday
         if familymember.birthday.strftime("%B %-d") == DateTime.now.strftime("%B %-d")
           flash[:info] = "Hey! Today is #{familymember.first_name}'s birthday. Wish #{familymember.first_name} a happy birthday on the Newsfeed!"
           p "Hey! Today is #{familymember.first_name}'s birthday. Wish them a happy birthday on your Newsfeed!"
-        elsif familymember.anniversary == DateTime.now
-          flash[:info] = "Hey! Today is #{familymember.first_name}'s anniversary. Wish #{familymember.first_name} a happy anniversary on the Newsfeed!"
         end
+      elsif familymember.anniversary == DateTime.now
+        flash[:info] = "Hey! Today is #{familymember.first_name}'s anniversary. Wish #{familymember.first_name} a happy anniversary on the Newsfeed!"
       end
-
-      render 'index.html.erb'
-    else
-      redirect_to '/login'
     end
+  
+    render "index.html.erb"
+
   end
 
   def new
@@ -50,9 +52,14 @@ class StatusUpdatesController < ApplicationController
 
   def destroy
     status_update = StatusUpdate.find(params[:id])
-    status_update.destroy
-    flash[:warning] = "Your status update has been deleted."
+
+    if current_user.id == status_update.user_id || current_user.admin
+      status_update.destroy
+      flash[:warning] = "Your status update has been deleted."
+    end
+
     redirect_to '/newsfeed'
+  
   end
 
 end
